Export Panel moved to own component
-Export Panel component integrated in Resume Editor
