+++ src/components/ResumeEditorV2.tsx
@@ -7,7 +7,7 @@
 // - Users can re-add deleted sections through the "Add Section" button
 // - Section order and deletion state are saved automatically
 
-import React, { useState, useEffect, useRef, useCallback } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import { useSession } from "next-auth/react";
 import { useRouter } from "next/navigation";
 import {
@@ -23,15 +23,6 @@ import {
   ListItemText,
   ListItemIcon,
   Divider,
-  Drawer,
-  Select,
-  MenuItem,
-  FormControl,
-  InputLabel,
-  Slider,
-  FormControlLabel,
-  Stack,
-  Tooltip,
   Dialog,
   DialogTitle,
   DialogContent,
@@ -42,15 +33,10 @@ import {
   Delete as DeleteIcon,
   DeleteOutline as DeleteOutlineIcon,
   Close as CloseIcon,
-  Star as StarIcon,
   List as ListIcon,
-
-  Info as InfoIcon,
-  Check as CheckIcon,
   DragIndicator as DragIndicatorIcon,
-  RestartAlt as RestartAltIcon,
 } from "@mui/icons-material";
-import { ToggleButton } from "@mui/material";
+
 
 
 import { DragDropContext, Droppable, Draggable, DropResult, DragStart, DragUpdate } from '@hello-pangea/dnd';
@@ -74,6 +60,7 @@ import { PublicationsSection } from "./ResumeEditor/components/sections/Publicat
 import { AwardsSection } from "./ResumeEditor/components/sections/AwardsSection";
 import { VolunteerExperienceSection } from "./ResumeEditor/components/sections/VolunteerExperienceSection";
 import { ReferencesSection } from "./ResumeEditor/components/sections/ReferencesSection";
+import { ExportPanel } from "./ResumeEditor/components/ExportPanel";
 
 
 
@@ -258,18 +245,7 @@ export default function ResumeEditorV2({
     pageHeight: 1100,
   });
 
-  // PDF Preview state
-  const [previewHtml, setPreviewHtml] = useState<string>('');
-  const [previewLoading, setPreviewLoading] = useState(false);
-  const [previewError, setPreviewError] = useState<string>('');
 
-  // Cache for preview results to avoid unnecessary API calls
-  const previewCache = useRef<Map<string, string>>(new Map());
-
-  // Create a cache key from export settings
-  const getCacheKey = useCallback((): string => {
-    return JSON.stringify(exportSettings);
-  }, [exportSettings]);
 
   // Autosave state
 
@@ -302,75 +278,7 @@ export default function ResumeEditorV2({
 
 
 
-  // Load PDF preview when export panel opens or settings change
-  useEffect(() => {
-    if (exportPanelOpen && resumeId) {
-      // Create AbortController for request cancellation
-      const abortController = new AbortController();
 
-      // Check cache first
-      const cacheKey = getCacheKey();
-      const cachedHtml = previewCache.current.get(cacheKey);
-
-      if (cachedHtml) {
-        setPreviewHtml(cachedHtml);
-        setPreviewLoading(false);
-        setPreviewError('');
-        return;
-      }
-
-      // Debounce the API call to avoid excessive requests
-      const timeoutId = setTimeout(async () => {
-        try {
-          setPreviewLoading(true);
-          setPreviewError('');
-          const response = await fetch(`/api/resumes/${resumeId}/pdf-html`, {
-            method: 'POST',
-            headers: {
-              'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-              exportSettings: exportSettings
-            }),
-            signal: abortController.signal,
-          });
-
-          if (!response.ok) {
-            throw new Error('Failed to load preview');
-          }
-
-          const htmlContent = await response.text();
-
-          // Cache the result
-          previewCache.current.set(cacheKey, htmlContent);
-
-          // Limit cache size to prevent memory issues
-          if (previewCache.current.size > 20) {
-            const firstKey = previewCache.current.keys().next().value;
-            if (firstKey) {
-              previewCache.current.delete(firstKey);
-            }
-          }
-
-          setPreviewHtml(htmlContent);
-        } catch (err) {
-          if (err instanceof Error && err.name === 'AbortError') {
-            // Request was cancelled, don't show error
-            return;
-          }
-          console.error('Error loading preview:', err);
-          setPreviewError('Failed to load preview');
-        } finally {
-          setPreviewLoading(false);
-        }
-      }, 150); // Reduced to 150ms for faster response
-
-      return () => {
-        clearTimeout(timeoutId);
-        abortController.abort(); // Cancel any pending request
-      };
-    }
-  }, [exportPanelOpen, resumeId, getCacheKey, exportSettings]);
 
   const [sectionOrder, setSectionOrder] = useState([
     "Personal Info",
@@ -1219,7 +1127,7 @@ export default function ResumeEditorV2({
   }
 
   // Helper: map section titles to render functions
-  const SECTION_COMPONENTS: Record<string, (resumeData: ResumeData, setResumeData: React.Dispatch<React.SetStateAction<ResumeData>>) => JSX.Element> = {
+  const SECTION_COMPONENTS: Record<string, () => JSX.Element> = {
     "Personal Info": () => (
       <PersonalInfoSection
         profileData={profileData}
@@ -1401,41 +1309,7 @@ export default function ResumeEditorV2({
 
 
 
-  const handleResetFormatting = () => {
-    const currentTemplate = exportSettings.template;
 
-    if (currentTemplate === 'standard') {
-      setExportSettings(prev => ({
-        ...prev,
-        fontFamily: 'Times New Roman',
-        nameSize: 40,
-        sectionHeadersSize: 14,
-        subHeadersSize: 10.5,
-        bodyTextSize: 11,
-        sectionSpacing: 12,
-        entrySpacing: 9,
-        lineSpacing: 12,
-        topBottomMargin: 33,
-        sideMargins: 33,
-        alignTextLeftRight: false,
-      }));
-    } else if (currentTemplate === 'compact') {
-      setExportSettings(prev => ({
-        ...prev,
-        fontFamily: 'Times New Roman',
-        nameSize: 36,
-        sectionHeadersSize: 12,
-        subHeadersSize: 10,
-        bodyTextSize: 10,
-        sectionSpacing: 10,
-        entrySpacing: 6,
-        lineSpacing: 10,
-        topBottomMargin: 25,
-        sideMargins: 25,
-        alignTextLeftRight: false,
-      }));
-    }
-  };
 
   const handleDeleteResume = async () => {
     if (!resumeId) {
@@ -1587,7 +1461,7 @@ export default function ResumeEditorV2({
                           {/* Section content */}
                           <Box sx={{ flex: 1, pl: 3 }}>
                             {SECTION_COMPONENTS[section]
-                              ? SECTION_COMPONENTS[section](resumeData, setResumeData)
+                              ? SECTION_COMPONENTS[section]()
                               : (
                                 <Box sx={{ py: 2, textAlign: "center" }}>
                                   <Typography color="text.secondary">
@@ -1633,7 +1507,7 @@ export default function ResumeEditorV2({
                               {/* Section content */}
                               <Box sx={{ flex: 1 }}>
                                 {SECTION_COMPONENTS[section]
-                                  ? SECTION_COMPONENTS[section](resumeData, setResumeData)
+                                  ? SECTION_COMPONENTS[section]()
                                   : (
                                     <Box sx={{ py: 2, textAlign: "center" }}>
                                       <Typography color="text.secondary">
@@ -2080,1057 +1954,17 @@ export default function ResumeEditorV2({
         </>
       )}
 
-      {/* Export Resume Panel - Slides out from right */}
-      <Drawer
-        anchor="right"
+      <ExportPanel
         open={exportPanelOpen}
         onClose={handleExportClose}
-        onTransitionEnd={() => {
-          if (!exportPanelOpen) {
-            setExportPanelFullyClosed(true);
-            // Clear the fallback timeout if it exists
-            if (exportPanelFallbackTimeoutRef.current) {
-              clearTimeout(exportPanelFallbackTimeoutRef.current);
-              exportPanelFallbackTimeoutRef.current = null;
-            }
-          }
-        }}
-        sx={{
-          '& .MuiDrawer-paper': {
-            width: 1170,
-            backgroundColor: 'white',
-            borderLeft: '1px solid #e0e0e0',
-            boxShadow: '-2px 0 8px rgba(0,0,0,0.1)',
-            borderTopLeftRadius: 20,
-            borderBottomLeftRadius: 20,
-          },
-        }}
-      >
-        <Box sx={{ py: 2, px: 3, height: '100%', display: 'flex', flexDirection: 'column' }}>
-          {/* Header */}
-          <Box sx={{
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'space-between',
-            mb: 3,
-            pb: 2,
-            borderBottom: '1px solid #e0e0e0'
-          }}>
-            <Typography variant="h6" fontWeight={600}>
-              Export Resume
-            </Typography>
-            <IconButton onClick={handleExportClose} size="small">
-              <CloseIcon />
-            </IconButton>
-          </Box>
-
-          {/* Content - Two Column Layout */}
-          <Box sx={{ flex: 1, display: 'flex', gap: 3, overflow: 'hidden' }}>
-            {/* Left Column - Resume Preview */}
-            <Box sx={{
-              flex: 1,
-              overflowY: 'auto',
-              overflowX: 'hidden',
-              pr: 1,
-              // Webkit scrollbar styling - more specific
-              '&::-webkit-scrollbar': {
-                width: '8px !important',
-                backgroundColor: 'transparent !important',
-              },
-              '&::-webkit-scrollbar-track': {
-                background: 'transparent !important',
-                backgroundColor: 'transparent !important',
-              },
-              '&::-webkit-scrollbar-thumb': {
-                background: '#cccccc !important',
-                backgroundColor: '#cccccc !important',
-                borderRadius: '4px !important',
-                border: 'none !important',
-              },
-              '&::-webkit-scrollbar-thumb:hover': {
-                background: '#aaaaaa !important',
-                backgroundColor: '#aaaaaa !important',
-              },
-              '&::-webkit-scrollbar-button': {
-                display: 'none !important',
-              },
-              '&::-webkit-scrollbar-corner': {
-                background: 'transparent !important',
-              },
-              // Firefox scrollbar styling
-              scrollbarWidth: 'thin',
-              scrollbarColor: '#cccccc transparent',
-            }}>
-
-
-
-
-              {/* Resume Preview */}
-              <Box sx={{
-                overflowY: 'visible',
-                overflowX: 'hidden',
-                height: '100%',
-                display: 'flex',
-                justifyContent: 'center',
-                alignItems: 'flex-start',
-                pb: 2,
-              }}>
-                {previewLoading ? (
-                  <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100%' }}>
-                    <CircularProgress />
-                  </Box>
-                ) : previewError ? (
-                  <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100%' }}>
-                    <Alert severity="error">{previewError}</Alert>
-                  </Box>
-                ) : previewHtml ? (
-                  <Box sx={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    alignItems: 'center',
-                    transform: 'scale(0.80)',
-                    transformOrigin: 'top left',
-                    minHeight: 'fit-content',
-                    width: '100%',
-                  }}>
-                    <div
-                      dangerouslySetInnerHTML={{ __html: previewHtml }}
-                      style={{
-                        width: '100%',
-                        margin: 0,
-                        padding: 0,
-                      }}
-                    />
-                  </Box>
-                ) : null}
-              </Box>
-            </Box>
-
-            {/* Right Column - Resume Template Settings */}
-            <Box sx={{
-              width: '35%',
-              overflowY: 'auto',
-              px: 0.5,
-              // Webkit scrollbar styling
-              '&::-webkit-scrollbar': {
-                width: '10px',
-              },
-              '&::-webkit-scrollbar-track': {
-                background: 'transparent !important',
-                marginLeft: '4px',
-              },
-              '&::-webkit-scrollbar-thumb': {
-                background: '#cccccc !important',
-                borderRadius: '5px',
-              },
-              '&::-webkit-scrollbar-thumb:hover': {
-                background: '#aaaaaa !important',
-              },
-              '&::-webkit-scrollbar-button': {
-                display: 'block',
-                height: '8px',
-                border: 'none',
-              },
-              '&::-webkit-scrollbar-button:hover': {
-                // No background
-              },
-              // Firefox scrollbar styling
-              scrollbarWidth: 'thin',
-              scrollbarColor: '#cccccc transparent',
-            }}>
-              {/* Resume Template Section */}
-              <Box sx={{
-                mb: 4,
-                border: '1px solid #e0e0e0',
-                borderRadius: 2,
-                p: 2,
-                backgroundColor: 'white'
-              }}>
-                <Typography variant="subtitle1" fontWeight={600} sx={{ mb: 2 }}>
-                  Resume Template
-                </Typography>
-                <FormControl fullWidth sx={{ mb: 2 }}>
-                  <Select
-                    value={exportSettings.pageSize}
-                    onChange={(e) => {
-                      const newPageSize = e.target.value;
-                      setExportSettings(prev => ({
-                        ...prev,
-                        pageSize: newPageSize,
-                        pageWidth: newPageSize === 'letter' ? 850 : 794,
-                        pageHeight: newPageSize === 'letter' ? 1100 : 1123,
-                      }));
-                    }}
-                    sx={{
-                      height: 33,
-                      fontSize: 14,
-                      backgroundColor: '#f5f5f5',
-                      '& .MuiOutlinedInput-root': {
-                        backgroundColor: '#f5f5f5',
-                        '& fieldset': {
-                          border: 'none',
-                        },
-                        '&:hover fieldset': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused fieldset': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused': {
-                          border: `2px solid ${COLORS.primary}`,
-                          outline: 'none',
-                        },
-                      },
-                      '& .MuiOutlinedInput-notchedOutline': {
-                        border: 'none',
-                      },
-                      '&:hover .MuiOutlinedInput-notchedOutline': {
-                        border: 'none',
-                      },
-                      '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                        border: `2px solid ${COLORS.primary} !important`,
-                      },
-                      '& .MuiMenuItem-root.Mui-selected': {
-                        backgroundColor: COLORS.primary,
-                        color: 'white',
-                        '&:hover': {
-                          backgroundColor: COLORS.hover,
-                        },
-                      },
-                    }}
-                  >
-                    <MenuItem value="letter" sx={{ fontSize: 14 }}>Letter (8.5&quot; x 11&quot;)</MenuItem>
-                    <MenuItem value="a4" sx={{ fontSize: 14 }}>A4 (210mm x 297mm)</MenuItem>
-                  </Select>
-                </FormControl>
-
-                {/* Template Previews */}
-                <Box sx={{ display: 'flex', gap: 2 }}>
-                  <Box sx={{
-                    flex: 1,
-                    height: 80,
-                    border: '2px solid #e0e0e0',
-                    borderRadius: 2,
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    backgroundColor: 'white',
-                    position: 'relative',
-                    cursor: 'pointer',
-                    ...(exportSettings.template === 'standard' && {
-                      borderColor: COLORS.primary,
-                      backgroundColor: COLORS.selectedLightGray,
-                      color: 'black',
-                    })
-                  }}
-                    onClick={() => setExportSettings(prev => ({
-                      ...prev,
-                      template: 'standard',
-                      fontFamily: 'Times New Roman',
-                      nameSize: 40,
-                      sectionHeadersSize: 14,
-                      subHeadersSize: 10.5,
-                      bodyTextSize: 11,
-                      sectionSpacing: 12,
-                      entrySpacing: 9,
-                      lineSpacing: 12,
-                      topBottomMargin: 33,
-                      sideMargins: 33,
-                      alignTextLeftRight: false
-                    }))}
-                  >
-                    <Typography variant="body2" fontWeight={500}>Standard</Typography>
-                    {exportSettings.template === 'standard' && (
-                      <Box sx={{ position: 'absolute', top: 4, right: 4 }}>
-                        <CheckIcon sx={{ color: COLORS.primary, fontSize: 16 }} />
-                      </Box>
-                    )}
-                    <Box sx={{ position: 'absolute', bottom: 4, right: 4 }}>
-                      <StarIcon sx={{ color: '#ffd700', fontSize: 14 }} />
-                    </Box>
-                  </Box>
-
-                  <Box sx={{
-                    flex: 1,
-                    height: 80,
-                    border: '2px solid #e0e0e0',
-                    borderRadius: 2,
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    backgroundColor: 'white',
-                    position: 'relative',
-                    cursor: 'pointer',
-                    ...(exportSettings.template === 'compact' && {
-                      borderColor: COLORS.primary,
-                      backgroundColor: COLORS.selectedLightGray,
-                      color: 'black',
-                    })
-                  }}
-                    onClick={() => setExportSettings(prev => ({
-                      ...prev,
-                      template: 'compact',
-                      fontFamily: 'Times New Roman',
-                      nameSize: 36,
-                      sectionHeadersSize: 12,
-                      subHeadersSize: 10,
-                      bodyTextSize: 10,
-                      sectionSpacing: 10,
-                      entrySpacing: 6,
-                      lineSpacing: 10,
-                      topBottomMargin: 25,
-                      sideMargins: 25,
-                      alignTextLeftRight: false
-                    }))}
-                  >
-                    <Typography variant="body2" fontWeight={500}>Compact</Typography>
-                    {exportSettings.template === 'compact' && (
-                      <Box sx={{ position: 'absolute', top: 4, right: 4 }}>
-                        <CheckIcon sx={{ color: COLORS.primary, fontSize: 16 }} />
-                      </Box>
-                    )}
-                  </Box>
-                </Box>
-
-                <Box sx={{ display: 'flex', alignItems: 'center', gap: 1, mt: 1 }}>
-                  <StarIcon sx={{ color: '#ffd700', fontSize: 14 }} />
-                  <Typography variant="caption" color="text.secondary">Recommended</Typography>
-                  <Tooltip title="This template is optimized for ATS (Applicant Tracking Systems) and follows professional resume standards" arrow>
-                    <InfoIcon sx={{ color: '#666', fontSize: 14 }} />
-                  </Tooltip>
-                </Box>
-              </Box>
-
-              {/* Font Section */}
-              <Box sx={{
-                mb: 4,
-                border: '1px solid #e0e0e0',
-                borderRadius: 2,
-                p: 2,
-                backgroundColor: 'white'
-              }}>
-                <Typography variant="subtitle1" fontWeight={600} sx={{ mb: 2 }}>
-                  Font
-                </Typography>
-
-                <FormControl fullWidth sx={{ mb: 2 }}>
-                  <InputLabel sx={{ position: 'static', transform: 'none', marginBottom: 1, fontSize: 14, fontWeight: 600, color: "black" }}>Font Family</InputLabel>
-                  <Select
-                    value={exportSettings.fontFamily}
-                    onChange={(e) => setExportSettings(prev => ({ ...prev, fontFamily: e.target.value }))}
-                    sx={{
-                      height: 33,
-                      fontSize: 14,
-                      backgroundColor: '#f5f5f5',
-                      '& .MuiOutlinedInput-root': {
-                        backgroundColor: '#f5f5f5',
-                        '& fieldset': {
-                          border: 'none',
-                        },
-                        '&:hover fieldset': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused fieldset': {
-                          border: `2px solid ${COLORS.primary} !important`,
-                        },
-                        '&.Mui-focused': {
-                          border: `2px solid ${COLORS.primary} !important`,
-                          outline: 'none',
-                        },
-                      },
-                      '& .MuiOutlinedInput-notchedOutline': {
-                        border: 'none',
-                      },
-                      '&:hover .MuiOutlinedInput-notchedOutline': {
-                        border: 'none',
-                      },
-                      '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                        border: `2px solid ${COLORS.primary} !important`,
-                      },
-                    }}
-                  >
-                    <MenuItem value="Times New Roman">Times New Roman</MenuItem>
-                    <MenuItem value="Arial">Arial</MenuItem>
-                    <MenuItem value="Calibri">Calibri</MenuItem>
-                    <MenuItem value="Georgia">Georgia</MenuItem>
-                    <MenuItem value="Helvetica">Helvetica</MenuItem>
-                    <MenuItem value="Verdana">Verdana</MenuItem>
-                    <MenuItem value="Garamond">Garamond</MenuItem>
-                    <MenuItem value="Bookman">Bookman</MenuItem>
-                  </Select>
-                </FormControl>
-
-                <Box sx={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: 3 }}>
-                  <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-                    <Typography variant="body2" sx={{ fontSize: 14, fontWeight: 600, color: "black" }}>Name</Typography>
-                    <Select
-                      value={exportSettings.nameSize}
-                      onChange={(e) => setExportSettings(prev => ({ ...prev, nameSize: Number(e.target.value) }))}
-                      sx={{
-                        height: 33,
-                        fontSize: 14,
-                        backgroundColor: '#f5f5f5',
-                        '& .MuiOutlinedInput-root': {
-                          backgroundColor: '#f5f5f5',
-                          '& fieldset': {
-                            border: 'none',
-                          },
-                          '&:hover fieldset': {
-                            border: 'none',
-                          },
-                          '&.Mui-focused fieldset': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                          },
-                          '&.Mui-focused': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                            outline: 'none',
-                          },
-                        },
-                        '& .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&:hover .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                          border: `2px solid ${COLORS.primary} !important`,
-                        },
-                      }}
-                      MenuProps={{
-                        PaperProps: {
-                          sx: {
-                            '& .MuiMenuItem-root': {
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                            '& .MuiMenuItem-root.Mui-selected': {
-                              backgroundColor: COLORS.selected,
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                            '&::-webkit-scrollbar': {
-                              width: '8px',
-                            },
-                            '&::-webkit-scrollbar-track': {
-                              background: 'transparent',
-                            },
-                            '&::-webkit-scrollbar-thumb': {
-                              background: '#ccc',
-                              borderRadius: '4px',
-                            },
-                            '&::-webkit-scrollbar-thumb:hover': {
-                              background: '#999',
-                            },
-                          },
-                        },
-                      }}
-                    >
-                      {[18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52].map((size) => (
-                        <MenuItem key={size} value={size} sx={{ fontSize: 14 }}>{size}</MenuItem>
-                      ))}
-                    </Select>
-                  </Box>
-
-                  <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-                    <Typography variant="body2" sx={{ fontSize: 14, fontWeight: 600, color: "black" }}>Section Headers</Typography>
-                    <Select
-                      value={exportSettings.sectionHeadersSize}
-                      onChange={(e) => setExportSettings(prev => ({ ...prev, sectionHeadersSize: Number(e.target.value) }))}
-                      MenuProps={{
-                        PaperProps: {
-                          sx: {
-                            '& .MuiMenuItem-root': {
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                            '& .MuiMenuItem-root.Mui-selected': {
-                              backgroundColor: COLORS.selected,
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                          },
-                        },
-                      }}
-                      sx={{
-                        height: 33,
-                        fontSize: 14,
-                        backgroundColor: '#f5f5f5',
-                        '& .MuiOutlinedInput-root': {
-                          backgroundColor: '#f5f5f5',
-                          '& fieldset': {
-                            border: 'none',
-                          },
-                          '&:hover fieldset': {
-                            border: 'none',
-                          },
-                          '&.Mui-focused fieldset': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                          },
-                          '&.Mui-focused': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                            outline: 'none',
-                          },
-                        },
-                        '& .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&:hover .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                          border: `2px solid ${COLORS.primary} !important`,
-                        },
-                      }}
-                    >
-                      {[10, 11, 12, 13, 14, 15].map((size) => (
-                        <MenuItem key={size} value={size} sx={{ fontSize: 14 }}>{size}</MenuItem>
-                      ))}
-                    </Select>
-                  </Box>
-
-                  <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-                    <Box sx={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                      <Typography variant="body2" sx={{ fontSize: 14, fontWeight: 600, color: "black" }}>Sub-Headers</Typography>
-                      <Tooltip title="Font size for job titles, company names, and other sub-headers in your resume" arrow>
-                        <InfoIcon sx={{ color: '#666', fontSize: 16 }} />
-                      </Tooltip>
-                    </Box>
-                    <Select
-                      value={exportSettings.subHeadersSize}
-                      onChange={(e) => setExportSettings(prev => ({ ...prev, subHeadersSize: Number(e.target.value) }))}
-                      MenuProps={{
-                        PaperProps: {
-                          sx: {
-                            '& .MuiMenuItem-root': {
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                            '& .MuiMenuItem-root.Mui-selected': {
-                              backgroundColor: COLORS.selected,
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                          },
-                        },
-                      }}
-                      sx={{
-                        height: 33,
-                        fontSize: 14,
-                        backgroundColor: '#f5f5f5',
-                        '& .MuiOutlinedInput-root': {
-                          backgroundColor: '#f5f5f5',
-                          '& fieldset': {
-                            border: 'none',
-                          },
-                          '&:hover fieldset': {
-                            border: 'none',
-                          },
-                          '&.Mui-focused fieldset': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                          },
-                          '&.Mui-focused': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                            outline: 'none',
-                          },
-                        },
-                        '& .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&:hover .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                          border: `2px solid ${COLORS.primary} !important`,
-                        },
-                      }}
-                    >
-                      {[8, 9, 10, 10.5, 11, 12, 13, 14].map((size) => (
-                        <MenuItem key={size} value={size} sx={{ fontSize: 14 }}>{size}</MenuItem>
-                      ))}
-                    </Select>
-                  </Box>
-
-                  <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-                    <Typography variant="body2" sx={{ fontSize: 14, fontWeight: 600, color: "black" }}>Body Text</Typography>
-                    <Select
-                      value={exportSettings.bodyTextSize}
-                      onChange={(e) => setExportSettings(prev => ({ ...prev, bodyTextSize: Number(e.target.value) }))}
-                      MenuProps={{
-                        PaperProps: {
-                          sx: {
-                            '& .MuiMenuItem-root': {
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                            '& .MuiMenuItem-root.Mui-selected': {
-                              backgroundColor: COLORS.selected,
-                              '&:hover': {
-                                backgroundColor: COLORS.selected,
-                              },
-                            },
-                          },
-                        },
-                      }}
-                      sx={{
-                        height: 33,
-                        fontSize: 14,
-                        backgroundColor: '#f5f5f5',
-                        '& .MuiOutlinedInput-root': {
-                          backgroundColor: '#f5f5f5',
-                          '& fieldset': {
-                            border: 'none',
-                          },
-                          '&:hover fieldset': {
-                            border: 'none',
-                          },
-                          '&.Mui-focused fieldset': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                          },
-                          '&.Mui-focused': {
-                            border: `2px solid ${COLORS.primary} !important`,
-                            outline: 'none',
-                          },
-                        },
-                        '& .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&:hover .MuiOutlinedInput-notchedOutline': {
-                          border: 'none',
-                        },
-                        '&.Mui-focused .MuiOutlinedInput-notchedOutline': {
-                          border: `2px solid ${COLORS.primary} !important`,
-                        },
-                      }}
-                    >
-                      {[8, 9, 10, 11, 12].map((size) => (
-                        <MenuItem key={size} value={size} sx={{ fontSize: 14 }}>{size}</MenuItem>
-                      ))}
-                    </Select>
-                  </Box>
-                </Box>
-              </Box>
-
-              {/* Spacing & Margin Section */}
-              <Box sx={{
-                mb: 4,
-                border: '1px solid #e0e0e0',
-                borderRadius: 2,
-                p: 2,
-                backgroundColor: 'white'
-              }}>
-                <Typography variant="subtitle1" fontWeight={600} sx={{ mb: 2 }}>
-                  Spacing & Margin
-                </Typography>
-
-                <Box sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
-                  <Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', mb: 1 }}>
-                      <Typography variant="body2">Section Spacing</Typography>
-                      <Tooltip title="Space between major sections like Work Experience, Education, Skills, etc." arrow>
-                        <InfoIcon sx={{ color: '#666', fontSize: 16 }} />
-                      </Tooltip>
-                    </Box>
-                    <Slider
-                      value={exportSettings.sectionSpacing}
-                      onChange={(_, value) => setExportSettings(prev => ({ ...prev, sectionSpacing: value as number }))}
-                      min={4}
-                      max={20}
-                      step={1}
-                      valueLabelDisplay="auto"
-                      valueLabelFormat={(value) => `${value} pt`}
-                      sx={{
-                        '& .MuiSlider-thumb': {
-                          backgroundColor: 'white',
-                          border: '1px solid black',
-                          transition: 'none',
-                          '&:hover': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:focus': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:active': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&::after': {
-                            content: '""',
-                            position: 'absolute',
-                            top: '50%',
-                            left: '50%',
-                            transform: 'translate(-50%, -50%)',
-                            width: '6px',
-                            height: '6px',
-                            backgroundColor: 'black',
-                            borderRadius: '50%',
-                          },
-                        },
-                        '& .MuiSlider-track': {
-                          backgroundColor: COLORS.primary,
-                          border: 'none',
-                        },
-                        '& .MuiSlider-rail': {
-                          backgroundColor: '#e0e0e0',
-                        },
-                        '& .MuiSlider-valueLabel': {
-                          borderRadius: '8px',
-                          padding: '4px 6px',
-                          backgroundColor: 'black',
-                        },
-                      }}
-                    />
-                  </Box>
-
-                  <Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', mb: 1 }}>
-                      <Typography variant="body2">Entry Spacing</Typography>
-                      <Tooltip title="Space between individual entries within a section (e.g., between different jobs in Work Experience)" arrow>
-                        <InfoIcon sx={{ color: '#666', fontSize: 16 }} />
-                      </Tooltip>
-                    </Box>
-                    <Slider
-                      value={exportSettings.entrySpacing}
-                      onChange={(_, value) => setExportSettings(prev => ({ ...prev, entrySpacing: value as number }))}
-                      min={4}
-                      max={16}
-                      step={1}
-                      valueLabelDisplay="auto"
-                      valueLabelFormat={(value) => `${value} pt`}
-                      sx={{
-                        '& .MuiSlider-thumb': {
-                          backgroundColor: 'white',
-                          border: '1px solid black',
-                          transition: 'none',
-                          '&:hover': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:focus': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:active': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&::after': {
-                            content: '""',
-                            position: 'absolute',
-                            top: '50%',
-                            left: '50%',
-                            transform: 'translate(-50%, -50%)',
-                            width: '6px',
-                            height: '6px',
-                            backgroundColor: 'black',
-                            borderRadius: '50%',
-                          },
-                        },
-                        '& .MuiSlider-track': {
-                          backgroundColor: COLORS.primary,
-                          border: 'none',
-                        },
-                        '& .MuiSlider-rail': {
-                          backgroundColor: '#e0e0e0',
-                        },
-                        '& .MuiSlider-valueLabel': {
-                          borderRadius: '8px',
-                          padding: '4px 6px',
-                          backgroundColor: 'black',
-                        },
-                      }}
-                    />
-                  </Box>
-
-                  <Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', mb: 1 }}>
-                      <Typography variant="body2">Line Spacing</Typography>
-                      <Tooltip title="Space between lines of text within bullet points and descriptions" arrow>
-                        <InfoIcon sx={{ color: '#666', fontSize: 16 }} />
-                      </Tooltip>
-                    </Box>
-                    <Slider
-                      value={exportSettings.lineSpacing}
-                      onChange={(_, value) => setExportSettings(prev => ({ ...prev, lineSpacing: value as number }))}
-                      min={10}
-                      max={15}
-                      step={1}
-                      valueLabelDisplay="auto"
-                      valueLabelFormat={(value) => `${value} pt`}
-                      sx={{
-                        '& .MuiSlider-thumb': {
-                          backgroundColor: 'white',
-                          border: '1px solid black',
-                          transition: 'none',
-                          '&:hover': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:focus': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:active': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&::after': {
-                            content: '""',
-                            position: 'absolute',
-                            top: '50%',
-                            left: '50%',
-                            transform: 'translate(-50%, -50%)',
-                            width: '6px',
-                            height: '6px',
-                            backgroundColor: 'black',
-                            borderRadius: '50%',
-                          },
-                        },
-                        '& .MuiSlider-track': {
-                          backgroundColor: COLORS.primary,
-                          border: 'none',
-                        },
-                        '& .MuiSlider-rail': {
-                          backgroundColor: '#e0e0e0',
-                        },
-                        '& .MuiSlider-valueLabel': {
-                          borderRadius: '8px',
-                          padding: '4px 6px',
-                          backgroundColor: 'black',
-                        },
-                      }}
-                    />
-                  </Box>
-
-                  <Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', mb: 1 }}>
-                      <Typography variant="body2">Top & Bottom Margin</Typography>
-                      <Tooltip title="Space from the top and bottom edges of each page" arrow>
-                        <InfoIcon sx={{ color: '#666', fontSize: 16 }} />
-                      </Tooltip>
-                    </Box>
-                    <Slider
-                      value={exportSettings.topBottomMargin}
-                      onChange={(_, value) => setExportSettings(prev => ({ ...prev, topBottomMargin: value as number }))}
-                      min={10}
-                      max={50}
-                      step={1}
-                      valueLabelDisplay="auto"
-                      valueLabelFormat={(value) => `${value} pt`}
-                      sx={{
-                        '& .MuiSlider-thumb': {
-                          backgroundColor: 'white',
-                          border: '1px solid black',
-                          transition: 'none',
-                          '&:hover': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:focus': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:active': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&::after': {
-                            content: '""',
-                            position: 'absolute',
-                            top: '50%',
-                            left: '50%',
-                            transform: 'translate(-50%, -50%)',
-                            width: '6px',
-                            height: '6px',
-                            backgroundColor: 'black',
-                            borderRadius: '50%',
-                          },
-                        },
-                        '& .MuiSlider-track': {
-                          backgroundColor: COLORS.primary,
-                          border: 'none',
-                        },
-                        '& .MuiSlider-rail': {
-                          backgroundColor: '#e0e0e0',
-                        },
-                        '& .MuiSlider-valueLabel': {
-                          borderRadius: '8px',
-                          padding: '4px 6px',
-                          backgroundColor: 'black',
-                        },
-                      }}
-                    />
-                  </Box>
-
-                  <Box>
-                    <Box sx={{ display: 'flex', alignItems: 'center', justifyContent: 'space-between', mb: 1 }}>
-                      <Typography variant="body2">Side Margins</Typography>
-                      <Tooltip title="Space from the left and right edges of each page" arrow>
-                        <InfoIcon sx={{ color: '#666', fontSize: 16 }} />
-                      </Tooltip>
-                    </Box>
-                    <Slider
-                      value={exportSettings.sideMargins}
-                      onChange={(_, value) => setExportSettings(prev => ({ ...prev, sideMargins: value as number }))}
-                      min={30}
-                      max={50}
-                      step={1}
-                      valueLabelDisplay="auto"
-                      valueLabelFormat={(value) => `${value} pt`}
-                      sx={{
-                        '& .MuiSlider-thumb': {
-                          backgroundColor: 'white',
-                          border: '1px solid black',
-                          transition: 'none',
-                          '&:hover': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:focus': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&:active': {
-                            backgroundColor: 'white',
-                            boxShadow: 'none',
-                          },
-                          '&::after': {
-                            content: '""',
-                            position: 'absolute',
-                            top: '50%',
-                            left: '50%',
-                            transform: 'translate(-50%, -50%)',
-                            width: '6px',
-                            height: '6px',
-                            backgroundColor: 'black',
-                            borderRadius: '50%',
-                          },
-                        },
-                        '& .MuiSlider-track': {
-                          backgroundColor: COLORS.primary,
-                          border: 'none',
-                        },
-                        '& .MuiSlider-rail': {
-                          backgroundColor: '#e0e0e0',
-                        },
-                        '& .MuiSlider-valueLabel': {
-                          borderRadius: '8px',
-                          padding: '4px 6px',
-                          backgroundColor: 'black',
-                        },
-                      }}
-                    />
-                  </Box>
-
-                  <FormControlLabel
-                    control={
-                      <ToggleButton
-                        value="align"
-                        selected={exportSettings.alignTextLeftRight}
-                        onChange={() => setExportSettings(prev => ({ ...prev, alignTextLeftRight: !prev.alignTextLeftRight }))}
-                        sx={{
-                          mr: -1,
-                          width: 50,
-                          height: 24,
-                          borderRadius: '20px',
-                          border: 'none',
-                          backgroundColor: exportSettings.alignTextLeftRight ? COLORS.primary : '#e0e0e0',
-                          '&.Mui-selected': {
-                            backgroundColor: COLORS.primary,
-                          },
-                          '&:hover': {
-                            backgroundColor: '#e0e0e0',
-                          },
-                          '&.Mui-selected:hover': {
-                            backgroundColor: COLORS.primaryLight,
-                          },
-                          position: 'relative',
-                          '&::after': {
-                            content: '""',
-                            position: 'absolute',
-                            top: '3px',
-                            left: exportSettings.alignTextLeftRight ? 'calc(100% - 21px)' : '3px',
-                            width: 18,
-                            height: 18,
-                            backgroundColor: 'white',
-                            borderRadius: '50%',
-                            transition: 'left 0.2s ease',
-                          },
-                        }}
-                      />
-                    }
-                    label="Align Text Left & Right"
-                    labelPlacement="start"
-                    sx={{
-                      justifyContent: 'space-between',
-                      width: '100%',
-                      ml: 0,
-                      '& .MuiFormControlLabel-label': {
-                        fontSize: '0.875rem'
-                      }
-                    }}
-                  />
-                </Box>
-              </Box>
-
-              {/* Reset Formatting Button */}
-              <Box sx={{ mb: 4 }}>
-                <Button
-                  variant="outlined"
-                  onClick={handleResetFormatting}
-                  fullWidth
-                  startIcon={<RestartAltIcon />}
-                  sx={{
-                    borderRadius: 2,
-                    border: '1px solid #e0e0e0',
-                    color: '#666',
-                    textTransform: 'none',
-                    py: 1,
-                  }}
-                >
-                  Reset formatting
-                </Button>
-              </Box>
-
-
-            </Box>
-          </Box>
-
-          {/* Footer - Download Buttons */}
-          <Stack direction="row" spacing={2} justifyContent="center" sx={{ pt: 2, borderTop: '1px solid #e0e0e0' }}>
-            <Button
-              variant="contained"
-              onClick={handleDownloadPDF}
-              disabled={pdfDownloading}
-              startIcon={pdfDownloading ? <CircularProgress size={16} color="inherit" /> : null}
-              sx={{
-                borderRadius: 6,
-                backgroundColor: '#000',
-                color: 'white',
-                textTransform: 'none',
-                fontSize: 16,
-                boxShadow: 'none',
-                '&:disabled': {
-                  backgroundColor: '#666',
-                  color: 'white',
-                },
-              }}
-            >
-              {pdfDownloading ? 'Generating PDF...' : 'Download by PDF'}
-            </Button>
-
-          </Stack>
-        </Box>
-      </Drawer>
-
+        exportSettings={exportSettings}
+        setExportSettings={setExportSettings}
+        resumeId={resumeId || ''}
+        resumeData={resumeData}
+        onDownloadPDF={handleDownloadPDF}
+        pdfDownloading={pdfDownloading}
+      />
+      
       {/* Delete Confirmation Dialog */}
       <Dialog
         open={deleteConfirmOpen}
